src/server/requirements.txt
alembic==1.3.2
aniso8601==8.0.0
bcrypt==3.1.7
cffi==1.13.2
chardet==3.0.4
Click==7.0
entrypoints==0.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
pexpect==4.7.0
ptyprocess==0.6.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.13.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
requests==2.21.0
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.4.1